+++ MODSRC/vazkii/botania/common/item/equipment/bauble/ItemIcePendant.java
@@ -51,9 +51,9 @@ public class ItemIcePendant extends ItemBauble {
 				tickIceRemovers(player);
 
 			if(!player.isSneaking() && !player.isInsideOfMaterial(Material.water)) {
-				int x = (int) MathHelper.floor_double(player.posX);
-				int y = (int) MathHelper.floor_double(player.posY - (player.isInWater() ? 0 : 1) - (player.worldObj.isRemote ? 1 : 0));
-				int z = (int) MathHelper.floor_double(player.posZ);
+				int x = MathHelper.floor_double(player.posX);
+				int y = MathHelper.floor_double(player.posY - (player.isInWater() ? 0 : 1) - (player.worldObj.isRemote ? 1 : 0));
+				int z = MathHelper.floor_double(player.posZ);
 
 				int range = 3;
 				for(int i = -range - 1; i < range; i++)
